pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beachwatch"
version = "0.1.4"
description = "A Python wrapper for the NSW Government's Beachwatch API."
authors = [{ name = "Ben Nour", email = "hello@ben-nour.com" }]
dependencies = ["requests~=2.32.3", "python-dateutil==2.9.0.post0"]

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">= 3.8"

[project.urls]
homepage = "https://github.com/ben-n93/beachwatch"
repository = "https://github.com/ben-n93/beachwatch"

[tool.setuptools]
packages = ["beachwatch"]
